Make VTTCue.size() from 'int' to 'double'
https://bugs.webkit.org/show_bug.cgi?id=260365
rdar://problem/114089775

Reviewed by Jer Noble.

This patch aligns WebKit with Gecko / Firefox, Blink / Chromium and Web-Spec [1][2].

[1] https://w3c.github.io/webvtt/#the-vttcue-interface
[2] https://www.w3.org/Bugs/Public/show_bug.cgi?id=19836

Spec Commit: https://github.com/w3c/webvtt/commit/64825911acce356bdd95c5ce8a8755f87f1d0702

In Web-Spec, `size` is double but in WebKit's implementation, it was using 'int',
so this PR just changes it to 'double'.

* Source/WebCore/html/track/VTTCue.h: Change 'size' to 'double' and also usage in 'setSize()' argument
* Source/WebCore/html/track/VTTCue.cpp:
(VTTCue::setSize): Update argument to 'double'
* LayoutTests/imported/w3c/web-platform-tests/webvtt/api/VTTCue/size.html: Add Test Case (manually)
* LayoutTests/imported/w3c/web-platform-tests/webvtt/api/VTTCue/size-expected.txt: Add Test Case Expectation
* LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/track/track-element/vtt-cue-float-precision-expected.txt: Rebaselined
* LayoutTests/media/track/track-webvtt-tc019-cue-size.html: Removed in favor of 'WPT' test
* LayoutTests/media/track/track-webvtt-tc019-cue-size-expected.txt: Ditto

Canonical link: https://commits.webkit.org/267041@main

+++ Source/WebCore/html/track/VTTCue.cpp
@@ -1,6 +1,6 @@
 /*
  * Copyright (C) 2011, 2013 Google Inc.  All rights reserved.
- * Copyright (C) 2011-2019 Apple Inc. All rights reserved.
+ * Copyright (C) 2011-2023 Apple Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are
@@ -412,7 +412,7 @@ void VTTCue::setPositionAlign(PositionAlignSetting positionAlignment)
     didChange();
 }
 
-ExceptionOr<void> VTTCue::setSize(int size)
+ExceptionOr<void> VTTCue::setSize(double size)
 {
     // http://www.whatwg.org/specs/web-apps/current-work/multipage/the-video-element.html#dom-texttrackcue-size
     // On setting, if the new value is negative or greater than 100, then throw an IndexSizeError
